+++ app/src/js/components/ListActions/ListActions.js
@@ -17,8 +17,8 @@ const ListActions = ({
   completedBulkActions
 }) => {
   const [actionsExpanded, setFiltersExpanded] = useState(false);
-  const hasActions = Array.isArray(bulkActions) && bulkActions.length > 0;
-  const hasGroupAction = hasActions && groupAction;
+  const hasBulkActions = Array.isArray(bulkActions) && bulkActions.length > 0;
+  const hasGroupAction = hasBulkActions && groupAction;
 
   function handleBulkActionSuccess(results, error) {
     if (typeof onBulkActionSuccess === 'function') {
@@ -87,13 +87,13 @@ const ListActions = ({
             </div>
           </Collapse>
         </div>
-        <Timer
-          noheader={!hasActions}
+        {action && <Timer
+          noheader={!hasBulkActions}
           dispatch={dispatch}
           action={action}
           config={queryConfig}
           reload={completedBulkActions}
-        />
+        />}
       </>
     );
   }
@@ -101,24 +101,24 @@ const ListActions = ({
   function renderActions() {
     return (
       <div className='list-actions'>
-        {hasActions && (
+        {hasBulkActions && (
           <div className='form--controls'>
             {listBulkActions()}
           </div>
         )}
-        <Timer
-          noheader={!hasActions}
+        {action && <Timer
+          noheader={!hasBulkActions}
           dispatch={dispatch}
           action={action}
           config={queryConfig}
           reload={completedBulkActions}
-        />
+        />}
       </div>
     );
   }
 
   return (
-    <div className={`list-action-wrapper${!hasActions || !children ? ' no-actions' : ''}`}>
+    <div className={`list-action-wrapper${!hasBulkActions || !children ? ' no-actions' : ''}`}>
       {children}
       {hasGroupAction && renderGroupActions()}
       {!hasGroupAction && renderActions()}